requirements.txt
streamlit>=1.28.0
python-docx>=0.8.11
Office365-REST-Python-Client>=2.4.2
pandas>=1.5.0
docx2txt>=0.8
olefile>=0.46